Do the change mentioned in last commit; extract separate UnicodeDecomposedForm classes that are used by both Decomposer and  DecomposingReadStream, so changes to the recursive mapping  don't have to be applied in two or more places; all decomposition tests are back to green.
Normal API (readStream unicodeCodePoints unicodeNF(K)D) is unaffected.
Setting form of DecompingReadStream is restricted to initialization.

Oh, and I changed the granularity of the UnicodeNFD subclassResponsiblity so there's still only one method to implement for each decomposed form. 
This means a bit of duplicate work in isNonStarter: where unicodeCharacterData is consulted twice; first to see if there's a decomposition, and then to see if non-decomposed codePoint isStarter not, but I think/hope this is acceptable.